Fix this issue unable to login to a standalone instance
Symptoms: when logging in to a standalone instance (vCloud 5.6), user
gets the following error message:
'''
Can't login: The request has invalid accept header. Supported API
versions are: [1.5, 5.1, 5.5, 5.6]
'''

It is because the get_service_type returns wrong value.
Because the standalone instance (vCloud 5.6) returns code 406 for
end point /api/iam/login. So the get_service_type return vca instead
standalone.

This fix just quick patch to work with standalone instance. However, the
status_code checking for VCA should be similar to VCHS a few lines
below. (maybe check if the return code is 401?). I'll leave it to someone
more familiar with VCA to change it.