[VCDA-689] Integrate generalized Jenkins build script (#263)
Update the Jenkins pipeline to allow more advanced functionality and reduce effort to 
integrate with Jenkins. 

The Jenkins pipeline now works from scripts that can run in the pipeline or in the user environment. The pipeline is also able to test remote repositories prior to accepting a pull request. There is a single pipeline only that can handle both automated build on commit
as well as manual override to point to remote repos. Minor changes made to run_* scripts
for examples and system_tests but no visible behavior changes for development use. 
See the support directory for more information.