[VCDA-762] Invalid Query will raise OperationNotSupportedException instead of silently returning empty list as result. (#288)
Whenever a typed query with an invalid type value or a type value that is hidden from the current user is specified, pyvcloud is unable to construct the query url and chooses to silently return an empty list as result set. This change list changes the behavior to raise an OperationNotSupportedException, as well as log the issue in log file.
As a fallout search_tests needed to be modified to address this behavioral change.

Also fixed a small typo in one of the Type Query type values (these type values are case sensitive).

Testing done : Ran all system tests and verified that none of them are failing after this change.