[VCDA-786] Fix failure during template uploads to catalog.
While uploading a template to a catalog we make multiple PUT calls in succession and write the data bits on the server. However while making the PUT calls we are specifying the wrong range where the data should be written. This can cause vCD to forcibly close the TCP connection, which will eventually lead to failed uploads.

This change list fixes the value we are sending in Content-Range header of these PUT calls.

Signed by: Aritra Sen aritra.iitkgp@gmail.com