Fix test cases
After changing the Enum from flufl, I think validating the value whether it is part of GatewayBackingConfig is failing.

For now, I have removed the validation because it is okay. Just that it will fail in vCD.

Testing Done: Executed the gateway_test class and it succeeded without any error.