Fix share catalog functionality for api v29.0 (#421)
The link to share catalog was added to /api/catalog/{id}'s body in vCD api v30.0. For api versions v29.0 and below, the link lives in the body of /api/admin/catalog/{id}. The current change list takes care of this dichotomy, and makes sure that the share catalog functionality keeps working no matter what api version is specified.