Fix issue configuring IP settings inEdgeGateway (#482)
When creating EdgeGateway for VDC it is possible to provide IP settings with a ip address. However SDK fail to configure the specified ip in tern edge gateway gets the next available ip in the subnet. Issue is related to comparison of netmask str and int. Solution is to cast the value to int and compare.

Fixes 481

Signed-off-by: Chaminda Divitotawela <cdivitotawela@gmail.com>