VP-2041: [PySDK]Consolidate VM
VP-2041: Consolidate VM

This CLN contains functionality of consolidating VM which is implemented
in method consolidate() of vm.py class. This functionality test requires
VM to be in POWERED OFF state.

Testing Done:
One test method as test_0071_consolidate() in vm_tests.py class to test
this functionality and it is getting executed successfully.