Include tenant context header when creating catalog (#709)
Added support for including the tenant context header (X-VMWARE-VCLOUD-TENANT-CONTEXT) when creating an org. This change allows a non-system-admin user to create a catalog in another org.

Adding this support included the client POST request having an added param: extra_headers. To generalize this change, this extra param was added to GET and DELETE client requests.