changing update_user to enable disabling a user
In addition to when all 3 variables are None, the conditional "is_enabled or role_name or password" will return False if is_enabled is False, indicating that I want to disable the user account. The put_resource needs to be called in order to disable the account, so the conditional needs to succeed in order for that to happen.